ai-mood-music-backend/requirements.txt
fastapi
uvicorn
requests
openai-agents
python-dotenv
langchain-google-genai